Better state machine interface
