+++ packages/state-machine/src/index.ts
@@ -1,4 +1,1 @@
-export * from "./connection";
-export * from "./fork";
-export * from "./sequence";
-export * from "./runner"
+export * from "./stateMachine";
